# docs/swagger.yaml
definitions:
...
  github_com_raids-lab_crater_internal_resputil.ErrorCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40102
    - 40103
    - 40104
    - 40105
    - 40106
    - 40301
    - 40302
    - 40401
    - 99999
    type: integer
    x-enum-varnames:
    - OK
    - InvalidRequest
    - TokenExpired
    - TokenInvalid
    - MustRegister
    - RegisterTimeout
    - RegisterNotFound
    - InvalidCredentials
    - UserNotAllowed
    - UserEmailNotVerified
    - ServiceSshdNotFound
    - NotSpecified
  github_com_raids-lab_crater_internal_resputil.Response-any:
    properties:
      code:
        $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.ErrorCode'
      data: {}
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_pkg_crclient_ClusterNodeDetail:
    properties:
      code:
        $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.ErrorCode'
      data:
        $ref: '#/definitions/github_com_raids-lab_crater_pkg_crclient.ClusterNodeDetail'
      msg:
        type: string
    type: object
  github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_pkg_crclient_GPUInfo:
    properties:
      code:
        $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.ErrorCode'
      data:
        $ref: '#/definitions/github_com_raids-lab_crater_pkg_crclient.GPUInfo'
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-internal_handler_NodeMark:
    properties:
      code:
        $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.ErrorCode'
      data:
        $ref: '#/definitions/internal_handler.NodeMark'
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-string:
    properties:
      code:
        $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.ErrorCode'
      data:
        type: string
      msg:
        type: string
    type: object
  github_com_raids-lab_crater_pkg_crclient.ClusterNodeDetail:
    properties:
      address:
        type: string
      arch:
        type: string
      containerRuntimeVersion:
        type: string
      gpuArch:
        type: string
      gpuCount:
        type: integer
      gpuMemory:
        type: string
      kubeletVersion:
        type: string
      name:
        type: string
      os:
        type: string
      osVersion:
        type: string
      role:
        type: string
      status:
        $ref: '#/definitions/v1.NodeConditionType'
      taint:
        type: string
      time:
        type: string
    type: object
  github_com_raids-lab_crater_pkg_crclient.GPUInfo:
    properties:
      cudaVersion:
        type: string
      gpuArch:
        type: string
      gpuCount:
        type: integer
      gpuDriver:
        type: string
      gpuMemory:
        type: string
      gpuProduct:
        type: string
      gpuUtil:
        additionalProperties:
          format: float32
          type: number
        type: object
      haveGPU:
        type: boolean
      name:
        type: string
      relateJobs:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
    type: object
...
  internal_handler.NodeAnnotation:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  internal_handler.NodeLabel:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  internal_handler.NodeMark:
    properties:
      annotations:
        items:
          $ref: '#/definitions/internal_handler.NodeAnnotation'
        type: array
      labels:
        items:
          $ref: '#/definitions/internal_handler.NodeLabel'
        type: array
      taints:
        items:
          $ref: '#/definitions/internal_handler.NodeTaint'
        type: array
    type: object
  internal_handler.NodeTaint:
    properties:
      effect:
        type: string
      key:
        type: string
      value:
        type: string
    type: object
...
  v1.NodeConditionType:
    enum:
    - Ready
    - MemoryPressure
    - DiskPressure
    - PIDPressure
    - NetworkUnavailable
    type: string
    x-enum-varnames:
    - NodeReady
    - NodeMemoryPressure
    - NodeDiskPressure
    - NodePIDPressure
    - NodeNetworkUnavailable
...
info:
  contact: {}
  description: This is the API server for Crater, a Multi-tenant AI Model Training
    Platform based on Kubernetes.
  title: Crater API
  version: 1.0.0
paths:
...
  /v1/nodes:
    get:
      consumes:
      - application/json
      description: kubectl + prometheus获取节点基本信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回值描述，注意这里返回Json字符串，swagger无法准确解析
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-string'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 获取节点基本信息
      tags:
      - Node
  /v1/nodes/{name}:
    get:
      consumes:
      - application/json
      description: kubectl + prometheus获取节点详细信息
      parameters:
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回值
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_pkg_crclient_ClusterNodeDetail'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 获取节点详细信息
      tags:
      - Node
    put:
      consumes:
      - application/json
      description: 介绍函数的主要实现逻辑
      parameters:
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回值
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-string'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 更新节点调度状态
      tags:
      - 接口对应的标签
  /v1/nodes/{name}/annotation:
    delete:
      consumes:
      - application/json
      description: 删除指定节点的注解
      parameters:
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      - description: 注解信息（只需要key）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.NodeAnnotation'
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除注解
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-string'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 删除节点注解
      tags:
      - Node
    post:
      consumes:
      - application/json
      description: 为指定节点添加注解
      parameters:
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      - description: 注解信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.NodeAnnotation'
      produces:
      - application/json
      responses:
        "200":
          description: 成功添加注解
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-string'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 添加节点注解
      tags:
      - Node
  /v1/nodes/{name}/gpu/:
    get:
      consumes:
      - application/json
      description: 查询prometheus获取GPU各节点的利用率
      parameters:
      - description: 节点名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回值描述
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_pkg_crclient_GPUInfo'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 获取GPU各节点的利用率
      tags:
      - Node
  /v1/nodes/{name}/label:
    delete:
      consumes:
      - application/json
      description: 删除指定节点的标签
      parameters:
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      - description: 标签信息（只需要key）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.NodeLabel'
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除标签
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-string'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 删除节点标签
      tags:
      - Node
    post:
      consumes:
      - application/json
      description: 为指定节点添加标签
      parameters:
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      - description: 标签信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.NodeLabel'
      produces:
      - application/json
      responses:
        "200":
          description: 成功添加标签
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-string'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 添加节点标签
      tags:
      - Node
  /v1/nodes/{name}/mark:
    get:
      consumes:
      - application/json
      description: 获取指定节点的Labels、Annotations和Taints信息
      parameters:
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回节点标记信息
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-internal_handler_NodeMark'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 获取节点标记信息
      tags:
      - Node
  /v1/nodes/{name}/pod/:
    get:
      consumes:
      - application/json
      description: kubectl + prometheus获取节点Pod信息
      parameters:
      - description: 节点名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回值描述
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 获取节点Pod信息
      tags:
      - Node
  /v1/nodes/{name}/taint:
    delete:
      consumes:
      - application/json
      description: 删除指定节点的污点
      parameters:
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      - description: 污点信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.NodeTaint'
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除污点
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-string'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 删除节点污点
      tags:
      - Node
    post:
      consumes:
      - application/json
      description: 为指定节点添加污点
      parameters:
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      - description: 污点信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.NodeTaint'
      produces:
      - application/json
      responses:
        "200":
          description: 成功添加污点
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-string'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 添加节点污点
      tags:
      - Node
...
securityDefinitions:
  Bearer:
    description: 访问 /login 并获取 TOKEN 后，填入 'Bearer ${TOKEN}' 以访问受保护的接口
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"